New Abilities, Sorted filesystem
I moved around some stuff in the file system to have it tidy and neat. I have also added a dash ability for the player character which allows them to quickly move out of the way of attacks, I have also added a frenzy ability allowing them to shoot faster for a short period of time.